+++ TablePro/Views/Main/Extensions/MainContentView+Setup.swift
@@ -50,6 +50,7 @@ extension MainContentView {
                 } else if let tabIndex = tabManager.selectedTabIndex {
                     coordinator.rebuildTableQuery(at: tabIndex)
                 }
+                await coordinator.rebuildSelectedTableQueryForHiddenColumnsIfNeeded()
             }
             if payload.skipAutoExecute {
                 _ = await schemaLoad
@@ -148,6 +149,7 @@ extension MainContentView {
                 if let tableName = firstTab.tableContext.tableName {
                     coordinator.restoreLastHiddenColumnsForTable(tableName)
                     coordinator.restoreFiltersForTable(tableName)
+                    await coordinator.rebuildSelectedTableQueryForHiddenColumnsIfNeeded()
                 }
                 if let session = DatabaseManager.shared.activeSessions[connection.id],
                     session.isConnected